Fixed bug.
